pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["apps*", "config*", "core*"]
exclude = [
    "tests*",
    "docs*",
    "venv*",
    "media*",
    "static*",
    "templates*",
    "locale*",
]

[project]
name = "aircnc-clone"
version = "0.1.0"
description = "Airbnb clone API built with Django REST Framework"
readme = "README.md"
authors = [{ name = "Vu Khanh Du", email = "khanhdu1211@gmail.com" }]
license = { text = "MIT" }
requires-python = ">=3.7"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Framework :: Django",
    "Framework :: Django :: 3.2",
    "Framework :: Django :: 4.0",
    "Framework :: Django :: 4.1",
    "Framework :: Django :: 4.2",
]

# Core dependencies (compatible with Python 3.7)
dependencies = [
    "Django>=3.2,<4.0",
    "djangorestframework>=3.14,<4.0",
    "psycopg2-binary>=2.9,<3.0",
    "python-decouple>=3.8,<4.0",
    "djangorestframework-simplejwt>=5.2,<6.0",
    "django-allauth>=0.50,<1.0",
    "drf-spectacular>=0.26,<1.0",
    "django-cors-headers>=3.0,<5.0",
    "django-filter>=23.0,<24.0",
    "Pillow>=8.0,<11.0",
    "redis>=4.0,<5.0",
    "django-redis>=5.0,<6.0",
    "celery>=5.0,<6.0",
    "django-extensions>=3.2,<4.0",
]

[project.optional-dependencies]
# Development dependencies (compatible with Python 3.7)
dev = [
    "django-debug-toolbar>=3.0,<4.0",
    "ipython>=7.0,<8.0",
    "black>=22.0,<24.0",
    "flake8>=5.0,<7.0",
    "isort>=5.0,<6.0",
    "pre-commit>=2.0,<4.0",
]

# Testing dependencies
test = [
    "pytest>=7.4,<8.0",
    "pytest-django>=4.5,<5.0",
    "factory-boy>=3.3,<4.0",
    "coverage>=7.2,<8.0",
    "pytest-cov>=4.1,<5.0",
]

# Documentation dependencies
docs = ["mkdocs>=1.5,<2.0", "mkdocs-material>=9.0,<10.0"]

# Production dependencies
prod = [
    "gunicorn>=20.1,<21.0",
    "whitenoise>=6.4,<7.0",
    "sentry-sdk>=1.32,<2.0",
    "django-storages>=1.13,<2.0",
]

# All development dependencies
all = ["aircnc-clone[dev,test,docs]"]

[project.urls]
Homepage = "https://github.com/dwvu0812/aircnc-clone"
Repository = "https://github.com/dwvu0812/aircnc-clone.git"
Documentation = "https://aircnc-clone.readthedocs.io/"
"Bug Tracker" = "https://github.com/dwvu0812/aircnc-clone/issues"

[project.scripts]
aircnc = "manage:main"

# Tool configurations
[tool.black]
line-length = 88
target-version = ['py37', 'py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_django = "django"
known_first_party = "apps,config,core"
sections = [
    "FUTURE",
    "STDLIB",
    "DJANGO",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER",
]

[tool.flake8]
max-line-length = 88
exclude = [
    ".git",
    "__pycache__",
    "*/migrations/*",
    "venv",
    "env",
    "build",
    "dist",
]
ignore = [
    "E203", # whitespace before ':'
    "E501", # line too long (handled by black)
    "W503", # line break before binary operator
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
addopts = [
    "--reuse-db",
    "--cov=apps",
    "--cov=config",
    "--cov=core",
    "--cov-report=html",
    "--cov-report=term-missing",
]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["apps", "config", "core"]
omit = [
    "*/migrations/*",
    "*/venv/*",
    "*/env/*",
    "manage.py",
    "*/settings/*",
    "*/tests/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
